Update for reading app context
To allow for the app to be rename but retain functionallity of the restart script, the restart scripts have been designed to take their parent directories into account and replace the removal path of "DeleteMeToRestart" to match the app name.

To make this easier to do, the "DeleteMeToRestart" file has been moved to the "bin" directory with the scripts.