Upgrade to version 2.0.0
Current versions of Splunk do not require this action and it the scripted input included in this version of the Splunk Forwarder Repair Kit would replace the datetime.xml which would cause integrity issues with the Splunk installations.

The scripts to perform this action have been retired and removed from the app.